expose a method to add default connectors and dataSources to models
