+++ packages/graphql/server/contextBuilder/createDataSources.ts
@@ -3,7 +3,7 @@
  *
  * As a default, will generate Mongoose data sources
  */
-import type { Connector } from "@vulcanjs/crud/server";
+import { VulcanGraphqlModelServer } from "..";
 import type { VulcanGenericDataSource } from "./typings";
 
 interface ModelDataSources {
@@ -35,28 +35,3 @@ export const createDataSources =
     }, {});
     return dataSources;
   };
-
-import { MongoDataSource } from "apollo-datasource-mongodb";
-import type { Model } from "mongoose";
-import { createMongooseConnector } from "@vulcanjs/mongo";
-import { VulcanGraphqlModelServer } from "..";
-/**
- * Create a mongoose data source
- * @param model
- * @param connector
- * @returns
- */
-const createMongooseDataSource = (
-  model: VulcanGraphqlModelServer,
-  connector: Connector
-) => {
-  const rawCollection = connector.getRawCollection();
-  if (!rawCollection) {
-    console.warn(`Model ${model.name} has no rawCollection in its connector. If it is not a Mongoose model, please
-    manually provide a dataSource in model.graphql options.`);
-    return undefined;
-  }
-  // TODO: check that it's a mongoose model?
-  const mongooseModel = rawCollection as unknown as Model<any>;
-  return new MongoDataSource(mongooseModel);
-};
