+++ packages/mongo-apollo/index.ts
@@ -1,2 +1,1 @@
-// Exports will be available at "@vulcanjs/mongo"
-export const addDefaultMongoConnector = "foo";
+export * from "./addDefaultMongoConnector";
